# elearning/e-smart-learn/app/docs/swagger.yaml
definitions:
  apperror.AppError:
    properties:
      fields:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
      type:
        $ref: '#/definitions/apperror.ErrorType'
    type: object
  apperror.ErrorType:
    enum:
    - not_found
    - bad_request
    - internal_server
    - unauthorized
    - forbidden
    - duplicate_entry
    - validation_error
    type: string
    x-enum-varnames:
    - NotFound
    - BadRequest
    - InternalServer
    - Unauthorized
    - Forbidden
    - DuplicateEntry
    - ValidationError
...
  consts.CourseStatus:
    enum:
    - draft
    - pending_review
    - published
    - rejected
    - archived
    type: string
    x-enum-varnames:
    - CourseDraft
    - CoursePending
    - CoursePublished
    - CourseRejected
    - CourseArchived
  dto.ApiResponse:
    properties:
      data: {}
      errors:
        items:
          $ref: '#/definitions/apperror.AppError'
        type: array
      metadata: {}
      path:
        type: string
      process_id:
        type: string
      request: {}
      status:
        $ref: '#/definitions/dto.ResponseStatus'
    type: object
...
  dto.CategoryResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CourseResponse:
    properties:
      average_rate:
        type: number
      category:
        $ref: '#/definitions/dto.CategoryResponse'
      created_at:
        type: string
      description:
        type: string
      duration:
        type: integer
      id:
        type: string
      image:
        type: string
      instructor:
        $ref: '#/definitions/dto.InstructorProfileResponse'
      is_purchased:
        type: boolean
      old_price:
        type: number
      price:
        type: number
      slug:
        type: string
      status:
        $ref: '#/definitions/consts.CourseStatus'
      title:
        type: string
      total_student:
        type: integer
      updated_at:
        type: string
    type: object
...
  dto.CreateInstructorProfileRequest:
    properties:
      bio:
        maxLength: 1000
        minLength: 3
        type: string
      category_id:
        type: string
      education:
        maxLength: 2000
        minLength: 3
        type: string
      instagram_url:
        type: string
      linkedin_url:
        type: string
      youtube_url:
        type: string
    required:
    - bio
    - education
    type: object
...
  dto.InstructorProfile:
    properties:
      balance:
        type: number
      bio:
        type: string
      category:
        $ref: '#/definitions/dto.CategoryResponse'
      certifications:
        items:
          type: string
        type: array
      cv_url:
        type: string
      education:
        type: string
      id:
        type: string
      instagram_url:
        type: string
      linkedin_url:
        type: string
      portfolio_url:
        type: string
      rating_avg:
        type: number
      status:
        type: string
      total_course:
        type: integer
      total_student:
        type: integer
      years_of_experience:
        type: integer
      youtube_url:
        type: string
    type: object
  dto.InstructorProfileResponse:
    properties:
      balance:
        type: number
      bio:
        type: string
      category:
        $ref: '#/definitions/dto.CategoryResponse'
      courses:
        items:
          $ref: '#/definitions/dto.CourseResponse'
        type: array
      created_at:
        type: string
      education:
        type: string
      id:
        type: string
      instagram_url:
        type: string
      linkedin_url:
        type: string
      rating_avg:
        type: number
      total_course:
        type: integer
      total_student:
        type: integer
      updated_at:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
      youtube_url:
        type: string
    type: object
...
  dto.Pagination:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      sortBy:
        type: string
      sortOrder:
        type: string
      total:
        type: integer
    type: object
...
  dto.ResponseStatus:
    properties:
      code:
        type: integer
      type:
        type: string
    type: object
...
  dto.RoleResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.UpdateInstructorProfileRequest:
    properties:
      bio:
        maxLength: 1000
        minLength: 3
        type: string
      category_id:
        type: string
      education:
        maxLength: 2000
        minLength: 3
        type: string
      instagram_url:
        type: string
      linkedin_url:
        type: string
      youtube_url:
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      address:
        type: string
      avatar:
        type: string
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      instructor_profile:
        $ref: '#/definitions/dto.InstructorProfile'
      name:
        type: string
      phone_number:
        type: string
      read_blog_ids:
        items:
          type: string
        type: array
      role:
        $ref: '#/definitions/dto.RoleResponse'
      total_course_in_progress:
        type: integer
      total_courses_enrolled:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/instructor-profiles:
    get:
      consumes:
      - application/json
      description: Retrieve a paginated list of instructor profiles with filtering,
        sorting, and pagination support
      parameters:
      - default: 10
        description: 'Number of items per page (default: 10, max: 100)'
        in: query
        name: limit
        type: integer
      - default: 0
        description: 'Number of items to skip (default: 0)'
        in: query
        name: offset
        type: integer
      - default: created_at
        description: 'Field to sort by (default: created_at)'
        in: query
        name: sortBy
        type: string
      - default: desc
        description: 'Sort order: asc or desc (default: desc)'
        enum:
        - asc
        - desc
        in: query
        name: sortOrder
        type: string
      - description: Filter instructor profiles by name (partial match)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.InstructorProfileResponse'
                  type: array
                metadata:
                  $ref: '#/definitions/dto.Pagination'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Get paginated list of instructor profiles
      tags:
      - instructor-profiles
    post:
      consumes:
      - application/json
      description: Create a new instructorProfile with the provided name and description.
        InstructorProfile name must be unique.
      parameters:
      - description: InstructorProfile create request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInstructorProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.InstructorProfileResponse'
              type: object
        "400":
          description: Invalid request payload or validation failed
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create a new instructorProfile
      tags:
      - instructor-profiles
  /api/v1/instructor-profiles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a instructorProfile by ID. Cannot delete a instructorProfile
        that has associated courses.
      parameters:
      - description: InstructorProfile ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: InstructorProfile not found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete a instructorProfile
      tags:
      - instructor-profiles
    get:
      consumes:
      - application/json
      description: Retrieve a single instructorProfile by its ID
      parameters:
      - description: InstructorProfile ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.InstructorProfileResponse'
              type: object
        "400":
          description: Invalid UUID format
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: InstructorProfile not found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Get instructor profile by ID
      tags:
      - instructor-profiles
    put:
      consumes:
      - application/json
      description: Update a instructorProfile by ID with the provided name and description.
        InstructorProfile name must be unique.
      parameters:
      - description: InstructorProfile ID (UUID format)
        in: path
        name: id
        required: true
        type: string
      - description: InstructorProfile update request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateInstructorProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.InstructorProfileResponse'
              type: object
        "400":
          description: Invalid UUID format or validation failed
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "404":
          description: InstructorProfile not found
          schema:
            $ref: '#/definitions/dto.ApiResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Update an existing instructorProfile
      tags:
      - instructor-profiles
...
  /api/v1/users/{id}/follow:
    delete:
      consumes:
      - application/json
      description: Unfollow another user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Unfollow a user
      tags:
      - follows
    post:
      consumes:
      - application/json
      description: Follow another user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      security:
      - BearerAuth: []
      summary: Follow a user
      tags:
      - follows
  /api/v1/users/{id}/followers:
    get:
      consumes:
      - application/json
      description: Get paginated list of followers of a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      - description: sortBy
        in: query
        name: sortBy
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Get followers by user ID
      tags:
      - follows
  /api/v1/users/{id}/followings:
    get:
      consumes:
      - application/json
      description: Get paginated list of users followed by a user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: offset
        in: query
        name: offset
        type: integer
      - description: sortBy
        in: query
        name: sortBy
        type: string
      - description: sortOrder
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse'
      summary: Get followings by user ID
      tags:
      - follows
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"